+++ packages/ui-patterns/PrivacySettings/index.tsx
@@ -4,8 +4,7 @@ import { handleResetTelemetry, LOCAL_STORAGE_KEYS } from 'common'
 import Link from 'next/link'
 import { PropsWithChildren, useEffect, useState } from 'react'
 import { Modal, Toggle } from 'ui'
-
-import { useConsentValue } from '../ConsentToast'
+import { useConsentValue } from '../shared/consent'
 
 export const PrivacySettings = ({
   children,
